requirements.txt
requests==2.32.3
python-dotenv==0.9.1
python-telegram-bot==13.7
sqlalchemy~=2.0.41

python-dateutil~=2.9.0.post0
helpers~=0.2.0